+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+* Add backtraces to logged exceptions.
 * Dump to a logfile when the server crashes.
 * Display runtime exceptions in the client.
 * Bug fixes: